fix: remove all network calls, revert to fixed fee estimate
