+++ libs/ledger-live-common/src/families/hedera/tests/utils.test.ts
@@ -1,7 +1,8 @@
 import BigNumber from "bignumber.js";
 import type { Account } from "@ledgerhq/types-live";
-import estimateMaxSpendable from "../js-estimateMaxSpendable";
-import network from "../../../network";
+import type { Transaction } from "../types";
+import { calculateAmount } from "../utils";
+
 
 // Balance is 1 Hbar
 const account: Account = {
@@ -52,27 +53,48 @@ const account: Account = {
   },
 };
 
-describe("js-estimateMaxSpendable", () => {
-  let estimatedFees = new BigNumber("150200").multipliedBy(2); // 0.001502 ℏ (as of 2023-03-14)
+const transaction: Transaction = {
+  family: "hedera",
+  amount: new BigNumber(1),
+  recipient: "",
+  useAllAmount: undefined,
+};
 
-  it("should return hedera price if available", async () => {
-    // If get hedera price works, use real estimate, otherwise fallback to hard coded
-    try {
-      const { data } = await network({
-        method: "GET",
-        url: "https://countervalues.live.ledger.com/latest/direct?pairs=hbar:usd",
-      });
-      estimatedFees = new BigNumber(10000).dividedBy(data[0]);
-    } catch {
-      console.error("Could not fetch Hedera price");
-    }
+describe("utils", () => {
+  let estimatedFees = new BigNumber("150200"); // 0.001502 ℏ (as of 2023-03-14)
+
+  test("calculateAmount transaction.useAllAmount = true", async () => {
+    transaction.useAllAmount = true;
+
+    const amount = account.balance.minus(estimatedFees.multipliedBy(2));
+    const totalSpent = amount.plus(estimatedFees);
+    const data = {
+      amount,
+      totalSpent,
+    };
+
+    const result = await calculateAmount({
+      account,
+      transaction,
+    });
+
+    expect(result).toEqual(data);
   });
 
-  test("estimateMaxSpendable", async () => {
-    const result = await estimateMaxSpendable({
+  test("calculateAmount transaction.useAllAmount = false", async () => {
+    transaction.useAllAmount = false;
+
+    const amount = transaction.amount;
+    const totalSpent = amount.plus(estimatedFees);
+    const data = {
+      amount,
+      totalSpent,
+    };
+
+    const result = await calculateAmount({
       account,
+      transaction,
     });
-    const data = account.balance.minus(estimatedFees);
 
     expect(result).toEqual(data);
   });
