+++ libs/ledger-live-common/src/families/hedera/utils.ts
@@ -2,24 +2,15 @@ import estimateMaxSpendable from "./js-estimateMaxSpendable";
 import BigNumber from "bignumber.js";
 import type { Account } from "@ledgerhq/types-live";
 import type { Transaction } from "./types";
-import network from "../../network";
 
+// NOTE: Hedera declares stable fees in USD
+//       We could get more accurate fee estimations using network calls
+//       to https://countervalues.live.ledger.com/latest/direct?pairs=hbar:usd
+//       However, this would create 4 points of failure in the integration tests.
+//       While viable, this could create undue pressure on CI for incoming PR's.
+//       > transfer fee is $0.0001 USD (10_000 * hedera_price to get the fee in tinybar)
 export const estimatedFeeSafetyRate = 2;
-
-export async function getEstimatedFees(): Promise<BigNumber> {
-  try {
-    const { data } = await network({
-      method: "GET",
-      url: "https://countervalues.live.ledger.com/latest/direct?pairs=hbar:usd",
-    });
-
-    return new BigNumber(10000).dividedBy(data[0]);
-  } catch {
-    // as fees are based on a currency conversion, we stay
-    // on the safe side here and double the estimate for "max spendable"
-    return new BigNumber("150200").multipliedBy(estimatedFeeSafetyRate); // 0.001502 ℏ (as of 2023-03-14)
-  }
-}
+export const estimatedFees = new BigNumber("150200"); // 0.001502 ℏ (as of 2023-03-14)
 
 export async function calculateAmount({
   account,
@@ -34,10 +25,10 @@ export async function calculateAmount({
   const amount = transaction.useAllAmount
     ? await estimateMaxSpendable({ account })
     : transaction.amount;
-
+    
   return {
     amount,
-    totalSpent: amount.plus(await getEstimatedFees()),
+    totalSpent: amount.plus(estimatedFees),
   };
 }
